Fix #105: expose hook for reflective dispatch.
