+++ core/src/main/java/feign/RequestTemplate.java
@@ -49,6 +49,10 @@ import static feign.Util.valuesOrEmpty;
  */
 public final class RequestTemplate implements Serializable {
 
+  interface Factory {
+    /** create a request template using args passed to a method invocation. */
+    RequestTemplate create(Object[] argv);
+  }
 
   private String method;
   /* final to encourage mutable use vs replacing the object. */
